fix: remove tracking of existing tables in `created_tables`
As raised in #122, the current usage of the table creation mechanism is
broken because of wrong book keeping of which tables have been created.
This has previously been implemented with a set `created_tables` than
was handed around different places that used or created tables. The
wrong book keeping lead to tables not existing when test queries would
run or tables already existing when they should be created.

This PR gets rid of the `created_tables` set all together. Instead,
those places responsible for creating tables is expected to do its own
book keeping in order to find out which tables need to be created (and
all already do that book keeping). This PR implements this logic
currently in such a way that existing tables with the same name as those
that should be created are first dropped and then recreated. It is
conceivable to be more optimistic and not do anything if the table
already exists; however, that mechanism breaks if two different tables
are loaded with the same name and is thus more brittle.

The PR also fixes `test_substrait_function_names.py`, which used an
additional DuckDB connection to prepare the Substrait plan used for test
but did not load the tables used in the plans. The PR thus adds the
logic to load those tables.

I have compared the list of failing tests before and after this change:
no test newly fails and 15 previously failing tests now pass. However,
there is a much larger number of tests affected; they now just fail
because the corresponding system under test fails rather than because
the test harness is broken.

This resolves #112.

Signed-off-by: Ingo Müller <ingomueller@google.com>