fix: add missing arguments to Producer.produce_substrait
Some producers had an additional parameter `validate` but the interface
and the IbisProducer as well as one call site were missing it. This PR
introduces that parameter in those places. In also changes the
formatting in one of the producers to be consistent with the others.

I have manually test: no test outcome changes. Several tests for the
IbisProducer reported a wrong number of arguments before and now fail
because the producer cannot handle the test cases.

Signed-off-by: Ingo Müller <ingomueller@google.com>