ci: set up GH action that enforces black formatting on changed lines
This is achieved via the Python package `darker`, which checks that all
changed lines are formatted with "black" style. The intent is to strike
a balance between enforcing a style but not doing one large commit that
only reformats everything, which messes up git history. After this CI
rule has run for some time, the (hopefully small) remaining part of the
code can still be reformatted with black in one commit and black be
enforced from that moment of time.

I have done a quick comparison of various formatters and black seems to
be the one that results in a formatting that is closest to the current
format, so enforcing that in CI seems like the best choice.

Signed-off-by: Ingo Müller <ingomueller@google.com>