ci: run all matrix jobs to the end even if others fail (#140)
PR #132 changed the definition of Github CI workflows to use the
`matrix` feature. If not changed explicitly, the different variants in
the matrix are cancelled as soon as the first variant fails. This
currently prevents us from knowing which of the systems pass CI. This PR
thus sets `fail-fast: false` in order to let all variants in the matrix
to run to completion.

Signed-off-by: Ingo Müller <ingomueller@google.com>