fix: deactivate diagnostics of `substrait-validator` correctly. (#141)
In several places, some diagnostics of the `substrait-validator` were
attempted to deactivate by specifying a minimum level that was higher
that the maximum level. This does not work anymore with newer versions
of the validator. If attempted, we get failed assertions with the
message `min <= max`. This PR instead sets both minimum and maximum
level to `info` in these cases, which resolves the broken assertions.

Signed-off-by: Ingo Müller <ingomueller@google.com>