ci: fix and improve build of Isthmus producer
This most have broken during one of the recent updates of the git
submodule. This PR downgrades the version of Java to the one used in CI
of `substrait-java`. It also changes the way Isthmus is build: using the
Github action to set up Gradle, which enables caching, and calling
`gradlew` directly rather than the wrapper script.

Signed-off-by: Ingo Müller <ingomueller@google.com>