ci: add workflow that ensures that support matrix is up-to-date (#145)
This PR adds a Github workflow that runs the tests for the support
matrix, parses the results, and compares that with the currently
checked-in version of the matrix, making CI fail if there is a
discrepancy. This ensures that the matrix is always up-to-date with the
current test results. The PR also updates the matrix.

Signed-off-by: Ingo Müller <ingomueller@google.com>