feat: sort rows consumer and producer matrices
This is a temporary commit that makes a trivial change to the scripts
parsing consumer and producer pytest data for the support matrices and
updates the CSV files of those matrices. The next commit in this PR will
make a highly non-trivial change that does not require rewriting the CSV
files, so having this commit helps to show the the next one makes no
semantic changes.

Signed-off-by: Ingo Müller <ingomueller@google.com>