feat: reimplement consumer and producer pytest parsing with pandas
This is the main commit of this PR, reimplementing the scripts that
parse the pytest CSV results and bring them into the desired shape for
the support matrices. Note that this commit does not change the CSV
files, giving confidence that the do exactly the same as the previous
scripts.

Signed-off-by: Ingo Müller <ingomueller@google.com>