feat: drop temporary reordering of columns
The previous commit made some column reordering to fit the (arbitrary)
column order of the previous script such that the CSV files remained
identical (which gives confidence in the correctness of the change).
However, that order isn't particularly meaningful, so this commit
removes the manual reordering, making the columns alphabetically
ordered. This implies a rewrite of the CSV file of the matrix but the
change is simple, so confidence in correctness should also be high.

Signed-off-by: Ingo Müller <ingomueller@google.com>