feat: convert `*_configs.py` to JSON data
This commit adds a script that converts the test definitions in the
`*_configs.py` files to JSON files. This will be used by the subsequent
commit, which will read the test definitions from the new JSON files.
The two parts are separated into to commits because the second commit
overwrites the files that the first one reads from.