feat: use separate test mechanism for Ibis producer
This PR decouples the Ibis producer test from the tests of the remaining
producers, which are all SQL based. Since Ibis' front-end is purely
Python-based and, thus, test cases consist of Python code rather than
SQL strings, there is actually little overlap between the two classes of
producers. Trying to handle the two together has led to somewhat
cumbersome code where the two cases had to be handled differently in
many places. Separating the two arguable leads to code that is easier to
understand and maintain. To model this distinction, the top-level
`Producer` interface does now neither assume SQL nor Ibis; instead, the
SQL-based producers inherit from the new `SQLProducer` interface, which
only knows about SQL, and `IbisProducer` only knows about Ibis
expressions.

With this PR, the Ibis producer tests are now regular pytest-based unit
test, i.e., each function constructing a particular Ibis expression is
now a pytest target. This is facilitated by a decorator that wraps these
expression factories into code that converts the resulting expression to
Substraits and validates the result. In particular, those factory
functions are now not part of test "configs" anymore, which are used to
parametrize `test_producer_*_functions` test targets. As a minor
downside, this means that the Ibis expressions aren't tested against the
`function_names` test; however, I argue that that is a kind of plan
validation that we should do as part of all producer tests and something
that the `substrait-validator` might already do (and better).