feat: read JSON-based test definition from `*_configs.py`
This commit reads the JSON-based test definititions from the JSON files
generated by the previous commit into the variables where they were
defined previsouly in the `*_configs.py` files. As a consequence, there
is (almost) no visible change to the outside world; further refactoring
will happen in subsequent PRs. A small change was necessary in the test
case parametrization logic, which relied on the order of the dictionary
keys in the test definitions, which has changed during the conversion.