ci: install Isthmus using actions in support matrix workflow (#157)
This adopts the same way of building and installing Isthmus and its
dependencies in the support matrix workflow as in the test workflow,
namely using GH actions. The latter had been updated at some point but
not the former. The former also used the wrong version of Java, which
now made CI fail. (No idea why it didn't break earlier.)

Signed-off-by: Ingo Müller <ingomueller@google.com>