feat: do TPC-H plan validation for Isthmus with `Producer`
This commit reuses the plan validation that is part of the `Producer`
interface in the tests that check plan validity for the TPC-H queries
for Isthmus. This leads to the same result in terms of validity;
however, the error class is different in the one error case, so the
outcome changes from one error to another for that query.

Signed-off-by: Ingo Müller <ingomueller@google.com>