feat: test for outcome in case of success in TPC-H validity tests (#166)
The tests ensuring plan validity only recorded the outcome in case of
failure, not in case of success. This means that changing a test case
from failure outcome to success outcome *without* updating the snapshots
did not lead to CI failure, which is against the intention of the CI
architecture. This PR adds snapshot tests also in case of valid plans.
This does not currently change any outcomes since none of the TPC-H
plans passes the validity test.

Signed-off-by: Ingo Müller <ingomueller@google.com>