feat: do TPC-H plan validation for Isthmus with `Producer` (#168)
This PR reuses the plan validation that is part of the `Producer`
interface in the tests that check plan validity for the TPC-H queries
for Isthmus. In the first commit, the changes the diagnostic
suppression of the direct usage of the `substrait-validator` to match
that of the Producer interface and records the change `*_outcome.txt`
files. In the second commit, the testing logic is change. That state
then leads to the same result in terms of validity as in the first
commit; however, the error class is different in the one error case,
so the outcome changes from one error to another for that query.

Signed-off-by: Ingo Müller <ingomueller@google.com>