feat: use file-based test case parametrization for TPC-H tests
This commit uses the JSON files from the previous commit to parametrize
the test cases related to TPC-H. This is the next step in making those
test instances of the consumer and producer tests that are also used to
test functions and relations. For now, the test logic and, in
particular, the snapshots are left as is in order to help us convince
ourselves that this PR is only a refactoring that does not change *what*
is being tested (and only changes *how* it is being tested).