feat: express TPC-H plan generation with existing infra
This PR uses `substrait_producer_sql_test` from `common.py` to express
the TPC-H test targets that regenerate the Substrait plans using
Isthmus. This replaces test-specific code with generic code and is,
thus, a steps towards reducing the complexity of the tests. As a
consequence, the paths of some snapshots change; however, note in the
corresponding git commit that all files are renamed without content
changes.

Signed-off-by: Ingo Müller <ingomueller@google.com>