feat: use file-based test case parametrization for TPC-H tests (#170)
This PR uses JSON-file-based test definitions for the TPC-H tests
similar to the "functional" tests. This is the next step in making
those test instances of the consumer and producer tests that are also
used to test functions and relations. For now, the test logic and, in
particular, the snapshots are left as is in order to help us convince
ourselves that this PR is only a refactoring that does not change what
is being tested (and only changes how it is being tested).

The PR is organized as three commits: The first commit uses as one-off
script to convert the definitions in the tpch_test_cases.py file to
JSON files. The second commit uses the JSON files from the previous
commit to parametrize the test cases related to TPC-H. The third commit
removes the old parametrization logic and the code-based test
definitions.

Signed-off-by: Ingo Müller <ingomueller@google.com>