feat: express TPC-H validation tests with existing infra (#173)
* feat: express TPC-H validation tests with existing infra

This PR uses the existing `substrait_producer_sql_test` test logic to
implement the test for plan validity of the TPC-H queries. This leads to
a moving of the outcome snapshots of these tests; however, the content
of these files is unchanged. Since the validity tests use the same test
definitions as the plan generation test targets and have, thus,
conflicting snapshot sames with those tests, we slightly change the test
name for the validity tests as a work-around, which leads to new plan
snapshots.

Signed-off-by: Ingo Müller <ingomueller@google.com>