feat: snapshot schema for consumer tests (#176)
This PR snapshots the result schema for consumer tests in addition to
the result data. This is currently already being done for the TPC-H
tests, so this PR is preparation for merging the TPC-H tests with the
remaining test infrastructure. This change changes the paths of the
result (data) snapshot files, adds snapshots for the result schema, and
changes the content of the "outcome" snapshot files, which now tracks
schema and data conformance separately.

Signed-off-by: Ingo Müller <ingomueller@google.com>