feat: make TPC-H snapshot paths more meaningful
This PR changes the path scheme for the snapshot files of the TPC-H
tests. Previously, the test definitions where read from a relative path
with only one folder, such that that folder and the file name where
interpreted as test "category" and "group", respectively. The latter
corresponded to the query number, leading to one folder per query, which
is more nesting than meaningful and differs from the other test
categories. This PR adds the folder "integration" to the relative path,
such that the snapshots reside under `integration/tpch/qXX-file.ext` (or
similar).

Signed-off-by: Ingo Müller <ingomueller@google.com>