feat: move Isthmus-specific code to `isthmus_producer.py`
Some of the Isthmus-specific code lived in top-level files called
`java_definitions.py` and `context.py`, which didn't convey a lot of
meaning and suggested more general usage that what was actually
happening. This PR moves the code directly to `isthmus_producer.py` (the
only place where it is used) and or to the `producer` directory in a
file `isthmus_java.py`, which indicates what it is used for.

Signed-off-by: Ingo Müller <ingomueller@google.com>