chore: update `substrait-java` git submodule to v0.47.0 (#210)
That version is the next release after the previously used version.

The PR also updates the snapshots of Isthmus. The only change is the
addition of `"offset": 0`. This is probably a consquence of the new
version of Substrait that `substrait-java` is using, which slightly
changed the definition of `FetchRel.offset`.

Signed-off-by: Ingo Müller <ingomueller@google.com>