chore: update `substrait-java` git submodule to next version  (#212)
This PRupdates the `substrait-java` git submodule to the next commit,
`f15cbabd`, which breaks some of the snapshot tests.

The PR also updates the plan snapshots of several tests of TPC-H queries
for Isthmus. The plans have, in parts, change substatntially and I have
not verified all changes in detail. However, the version of Calcite that
Isthmus uses has changed in this update, so major plan changes seem
reasonable.

Finally, the PR updates the outcome of one DuckDB consumer test, which
is now able to execute an updated plan produced by Isthmus that it wasn't
previously able to consume.

Signed-off-by: Ingo Müller <ingomueller@google.com>