chore: update `substrait-java` git submodule to latest version (#213)
This PR updates the `substrait-java` git submodule to its latest
version as of today, namely v0.50.0.

The PR also updates one plan snapshot of an Isthmus test due to the
version change in this PR. The plan changes the members of an
`intervalDaysToSeconds`, so I suppose it is due to
substrait-io/substrait-java#335, which fixes a bug related to that
message type.

Signed-off-by: Ingo Müller <ingomueller@google.com>