chore: update `substrait-java` git submodule to v0.51.0  (#217)
This is the latest version that runs with just one fix. The breaking
change until there is substrait-io/substrait-java#350, which changes
which types are used in some places. Consequently, some of the plans
generated by Isthmus change: in several occasions, the precision and/or
scale of decimal types changes and, in one instance, a new cast to a
decimal type with a different precision is inserted.

Signed-off-by: Ingo Müller <ingomueller@google.com>