chore(deps): update substrait-java to next (breaking) version (#243)
This PR updates the `substrait-java` submodule to the next version,
which breaks the tests, namely substrait-io/substrait-java@007accf. The
breakages are due to substrait-io/substrait-java#408, which changes the
index at which anchors start, which basically changes every plan that
Isthmus produces.

This PR also updates all snapshots. The only changes seem to be updates
of anchors and references to them.

Signed-off-by: Ingo Müller <ingomueller@google.com>